+++ CHANGELOG.md
@@ -17,11 +17,16 @@ The minor version will be incremented upon a breaking change and the patch versi
 ## [10.6.0]
 
 ### Features
-- Added support for block leaders from yellowstone-blocklist smart contract
-- Refactor `cluster_tpu_info.rs` and `grpc_geyser.rs`
+
+- Added 'IdentityFlusherWaitGroup' struct which function is to control when an identity is updated in Jet.
+- Added reset_identity method in RPC.
+- Added flush transactions while updating identity.
+- Added shutdown connection with Jet-Gateway while updating identity.
 
 ### Fixes
-- Added tests for blocklist from yellowstone-blocklist
+
+- Refactor `SendTransactionsPool`, `SendTransactionsPoolTask`, `RootedTransactions`, `QuicClient`, `ClusterTpuInfo`, `BlockhasQueue` 
+- Added integration test for `SendTransactionsPool`, `IdentityFlusherWaitGroup` and `RpcServer`.
 
 ## [10.5.0]